Remove case insensitive culture search
