+++ src/System.Private.CoreLib/shared/System/Runtime/Loader/AssemblyLoadContext.cs
@@ -555,6 +555,37 @@ namespace System.Runtime.Loader
                 }
             }
         }
+
+        private Assembly? ResolveSatelliteAssembly(AssemblyName assemblyName)
+        {
+            string? cultureName = assemblyName.CultureName;
+
+            if (cultureName == null || cultureName.Length == 0)
+                return null;
+
+            if (assemblyName.Name == null)
+                return null;
+
+            AssemblyName parentAssemblyName = new AssemblyName(assemblyName.Name);
+
+            Assembly? parentAssembly = LoadFromAssemblyName(parentAssemblyName);
+
+            if (parentAssembly == null)
+                return null;
+
+            // ResolveSatelliteAssembly should always be called on the ALC which loaded parentAssembly
+            Debug.Assert(this == GetLoadContext(parentAssembly));
+
+            string parentDirectory = Path.GetDirectoryName(parentAssembly.Location)!;
+
+            string assemblyPath = Path.Combine(parentDirectory, cultureName, $"{assemblyName.Name}.dll");
+            if (Internal.IO.File.InternalExists(assemblyPath))
+            {
+                return LoadFromAssemblyPath(assemblyPath);
+            }
+
+            return null;
+        }
     }
 
     internal sealed class DefaultAssemblyLoadContext : AssemblyLoadContext
